+++ test/integration/default/default_test.rb
@@ -1,0 +1,5 @@
+# # encoding: utf-8
+
+describe http('http://localhost') do
+  its('status') { should eq 200 }
+end